Tell user how to correct content of dropdown
First time users of Mugen will be unable to advance and will spend hours trying to determine what happened to the content of the foreign key dropdown if they choose not to use "name" as the name of the attribute to be displayed.  This change makes the both the problem and the solution explicit.